[HeartRateMonitor] - Migrated to 64 bit
- Sample was updated to 64 bit.

+++ HeartRateMonitor/HeartRateMonitorTableDataSource.cs
@@ -27,8 +27,8 @@
 using System;
 using System.Collections.Generic;
 
-using MonoMac.Foundation;
-using MonoMac.AppKit;
+using Foundation;
+using AppKit;
 
 namespace Xamarin.HeartMonitor
 {
@@ -41,24 +41,24 @@ namespace Xamarin.HeartMonitor
 			heartRateMonitors.Add (peripheral);
 		}
 
-		public override NSObject GetObjectValue (NSTableView tableView, NSTableColumn tableColumn, int row)
+		public override NSObject GetObjectValue (NSTableView tableView, NSTableColumn tableColumn, nint row)
 		{
-			var peripheral = heartRateMonitors [row];
+			var peripheral = heartRateMonitors [(int)row];
 			return new NSString (peripheral.Name);
 		}
 
-		public override int GetRowCount (NSTableView tableView)
+		public override nint GetRowCount (NSTableView tableView)
 		{
 			return heartRateMonitors.Count;
 		}
 		
-		public HeartRateMonitor GetHeartRateMonitor (int row)
+		public HeartRateMonitor GetHeartRateMonitor (nint row)
 		{
 			if (row < 0 || row >= heartRateMonitors.Count) {
 				return null;
 			}
 
-			return heartRateMonitors [row];
+			return heartRateMonitors [(int)row];
 		}
 	}
 }
